refactor: remove dependency on gorilla/mux
This little tool is (not yet) intended to have a fully featured REST
API, so there's no need for this dependency.

This change also removes the new upload handler started recently in
favor of the older one which wasn't actually tied at all to gorilla/mux
and already handles the file upload fairly well.